♻️단일책임을 위한 컴포넌트 1차 분리리
