+++ src/shared/ui/HighlightText.tsx
@@ -1,0 +1,21 @@
+/**
+ *
+ * @param text
+ * @param highlight
+ * @returns <span />
+ */
+
+// 하이라이트 함수 추가
+export const highlightText = (text: string, highlight: string) => {
+  if (!text) return null
+  if (!highlight.trim()) {
+    return <span>{text}</span>
+  }
+  const regex = new RegExp(`(${highlight})`, "gi")
+  const parts = text.split(regex)
+  return (
+    <span>
+      {parts.map((part, i) => (regex.test(part) ? <mark key={i}>{part}</mark> : <span key={i}>{part}</span>))}
+    </span>
+  )
+}
